replaced old sponsor page with new; changed erika's pfp
