+++ src/App.jsx
@@ -6,8 +6,7 @@ import Home from "./Pages/Home.jsx";
 import FAQ from "./Pages/FAQ.jsx";
 import MeetTheTeam from "./Pages/MeetTheTeam.jsx";
 import Schedule from "./Pages/Schedule.jsx";
-// import Sponsors from "./Pages/Sponsors.jsx";
-import SponsorsV2 from "./Pages/SponsorsV2.jsx";
+import Sponsors from "./Pages/Sponsors.jsx";
 import AboutUs from "./Pages/About.jsx";
 import BH2025 from "./Pages/BH2025.jsx";
 import Footer from "./Components/Footer.jsx";
@@ -24,7 +23,7 @@ function App() {
         <Route path="/About" element={<AboutUs />} />
         <Route path="/FAQ" element={<FAQ />} />
         <Route path="/Team" element={<MeetTheTeam />} />
-        <Route path="/Sponsors" element={<SponsorsV2 />} />
+        <Route path="/Sponsors" element={<Sponsors />} />
         <Route path="/Schedule" element={<Schedule />} />
         <Route path="/BH2025" element={<BH2025 />} />
         <Route path="/*" element={<Error />} />
